Get the app to compile with inlining
