+++ src/index.html
@@ -2,11 +2,6 @@
 <html>
   <head>
     <meta charset="UTF-8" />
-    <meta property="csp-nonce" content="<%= nonce %>" />
-    <meta
-      http-equiv="Content-Security-Policy"
-      content="<%= contentSecurityPolicy %>"
-    />
     <meta
       name="viewport"
       content="minimum-scale=1, initial-scale=1, width=device-width, shrink-to-fit=no"
