--- webpack/globalConstants.js
@@ -1,5 +1,0 @@
-const nonce = require("./nonce");
-
-module.exports = {
-  __NONCE__: JSON.stringify(nonce),
-};
