--- webpack/nonce.js
@@ -1,7 +1,0 @@
-const fs = require("fs");
-
-module.exports = fs
-  .readFileSync(`${__dirname}/../.nonce`, {
-    encoding: "utf8",
-  })
-  .trim();
